limit nokogiri version in Gemfile due to Vercel GLIBC version
avoid incurring in 'unsupported version of glibc' error on Vercel.
Vercel build system uses a precompiled version of nokogiri, but it
requires a glibc supported version on the system. Fixing the nokogiri
version to something the glibc version on Vercel can handle.

Full error below:
ERROR: It looks like you're trying to use Nokogiri as a precompiled native gem on a system
       with an unsupported version of glibc.
  /lib64/libc.so.6: version `GLIBC_2.28' not found (required by /vercel/path0/vendor/bundle/ruby/3.2.0/gems/nokogiri-1.18.2-x86_64-linux-gnu/lib/nokogiri/3.2/nokogiri.so) - /vercel/path0/vendor/bundle/ruby/3.2.0/gems/nokogiri-1.18.2-x86_64-linux-gnu/lib/nokogiri/3.2/nokogiri.so